Refactor files/packages
