+++ app/src/main/res/layout/fragment_browser.xml
@@ -41,7 +41,7 @@
         android:layout_marginTop="56dp"
         android:visibility="gone">
 
-        <com.cliqz.browser.freshtab.NewsView
+        <com.cliqz.browser.news.ui.NewsView
             android:id="@+id/newsView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
