--- freshtab/src/main/java/com/cliqz/browser/freshtab/data/source/NewsDataSource.kt
@@ -1,8 +1,0 @@
-package com.cliqz.browser.freshtab.data.source
-
-import com.cliqz.browser.freshtab.data.NewsItem
-import com.cliqz.browser.freshtab.data.Result
-
-interface NewsDataSource {
-    suspend fun getNews(): Result<List<NewsItem>>
-}
