--- freshtab/src/main/java/com/cliqz/browser/freshtab/data/source/NewsRepository.kt
@@ -1,9 +1,0 @@
-package com.cliqz.browser.freshtab.data.source
-
-import com.cliqz.browser.freshtab.data.NewsItem
-import com.cliqz.browser.freshtab.data.Result
-
-interface NewsRepository {
-
-    suspend fun getNews(forceUpdate: Boolean = false): Result<List<NewsItem>>
-}
