+++ freshtab/src/main/java/com/cliqz/browser/news/data/source/DefaultNewsRepository.kt
@@ -1,9 +1,9 @@
-package com.cliqz.browser.freshtab.data.source
+package com.cliqz.browser.news.data.source
 
-import com.cliqz.browser.freshtab.data.NewsItem
-import com.cliqz.browser.freshtab.data.Result
-import com.cliqz.browser.freshtab.data.Result.Success
-import com.cliqz.browser.freshtab.data.source.remote.NewsRemoteDataSource
+import com.cliqz.browser.news.data.NewsItem
+import com.cliqz.browser.news.data.Result
+import com.cliqz.browser.news.data.Result.Success
+import com.cliqz.browser.news.data.source.remote.NewsRemoteDataSource
 
 class DefaultNewsRepository(
     private val newsRemoteDataSource: NewsRemoteDataSource
